moved `SimAlertType` and `SimAlertLevel` existing enums into its own file and updated deps
