+++ motorlib/enums/simAlertType.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class SimAlertType(Enum):
+    """Types of sim alerts"""
+    GEOMETRY = 1
+    CONSTRAINT = 2
+    VALUE = 3
