+++ motorlib/grains/finocyl.py
@@ -2,9 +2,11 @@
 
 import numpy as np
 
+from ..enums.simAlertLevel import SimAlertLevel
+from ..enums.simAlertType import SimAlertType
 from ..grain import FmmGrain
 from ..properties import FloatProperty, IntProperty
-from ..simResult import SimAlert, SimAlertLevel, SimAlertType
+from ..simResult import SimAlert
 
 class Finocyl(FmmGrain):
     """A finocyl (fins on cylinder) grain has a circular core with a number of rectangular extensions that start at the
