+++ motorlib/grains/moonBurner.py
@@ -1,8 +1,9 @@
 """Moon burning grain submodule"""
-
+from ..enums.simAlertLevel import SimAlertLevel
+from ..enums.simAlertType import SimAlertType
 from ..grain import FmmGrain
 from ..properties import FloatProperty
-from ..simResult import SimAlert, SimAlertLevel, SimAlertType
+from ..simResult import SimAlert
 
 class MoonBurner(FmmGrain):
     """A moonburner is very similar to a BATES grain except the core is off center by a specified distance."""
